// LinkERP/ClientApp/src/app/models/inv/lbs-inv-vendor-price-scheme.ts
import { BaseEntity } from "../base-entity"
export class LBSINVVendorPriceScheme extends BaseEntity {
  CompanyID: any;
  VendorID: any;
  DealsType: any;
  ProductID: any;
  ProductDescription: any;
  Quantity: any;
  UOM: any;
  SupplierUnitPrice: any;
  DiscountType: any;
  DiscountValue: any;
  FreeDealsQuantity: any;
  FreeDealsQuantityRedeemed: any;
  FreeDealsUOM: any;
  ValidFrom: any;
  ValidTo: any;
  Status: any;
}